Split generic tests into it's own file.
